+++ internal/provider/provider.go
@@ -141,6 +141,7 @@ func (p *TruenasProvider) DataSources(ctx context.Context) []func() datasource.D
 	return []func() datasource.DataSource{
 		NewDatasetDataSource,
 		NewPoolDataSource,
+		NewVMGuestInfoDataSource,
 	}
 }